requirements.txt
pandas
azure-storage-blob